Updated Program
I have updated the program and fixed some of the issues that were occuring.
I have also updated the program so new features have been added such as userid.
I have also added a Scan Build ID Textbox, with this if we get 4 scanners on the line, we would be able to scan the pc as its going down the line and from that with Arturas help, could create a database table and could have the pc build id, and track who built what on each step from the user ids.
I am hoping to fix all the other sound issues later this week however is was not a priority and wanted to get all the important stuff that needed doing out the way.